feat: fractional Bortle display (B5.7 not B6) + smart online caching
- bortleClass changed from Int? to Double? on ImageEntry + FrameRecord
- Column shows B5.7 format (one decimal place)
- estimateOnline() returns Double for fractional precision
- refineBortleOnline(): one Supabase call per unique lat/lon, applies to
  all frames at same coordinates. Fire-and-forget after header enrichment.
- Uses REST API directly (no Edge Function dependency)
- UserDefaults cache per coordinate pair (never re-fetches same location)

+++ AstroTriage/Model/FrameRecord.swift
@@ -82,7 +82,7 @@ struct FrameRecord: Codable, FetchableRecord, PersistableRecord {
     var filterTrailingMultiplier: Double?
 
     // Light pollution + target clustering
-    var bortleClass: Int?           // Bortle 1-9 from BortleEstimator
+    var bortleClass: Double?        // Bortle 1.0-9.0 fractional from BortleEstimator
     var canonicalTarget: String?    // Normalized target name for grouping
 
     var wasDeleted: Int             // 0 or 1
